+++ src/main/java/org/jboss/com/sun/corba/se/impl/protocol/giopmsgheaders/CancelRequestMessage_1_0.java
@@ -25,50 +25,53 @@
 
 package org.jboss.com.sun.corba.se.impl.protocol.giopmsgheaders;
 
-
 /**
  * This implements the GIOP 1.0 CancelRequest header.
- *
+ * 
  * @author Ram Jeyaraman 05/14/2000
  */
 
-public final class CancelRequestMessage_1_0 extends Message_1_0
-        implements CancelRequestMessage {
+public final class CancelRequestMessage_1_0 extends Message_1_0 implements CancelRequestMessage
+{
 
     // Instance variables
 
-    private int request_id = (int) 0;
+    private int request_id = 0;
 
     // Constructors
 
-    CancelRequestMessage_1_0() {}
+    CancelRequestMessage_1_0()
+    {
+    }
 
-    CancelRequestMessage_1_0(int _request_id) {
-        super(Message.GIOPBigMagic, false, Message.GIOPCancelRequest,
-              CANCEL_REQ_MSG_SIZE);
+    CancelRequestMessage_1_0(int _request_id)
+    {
+        super(Message.GIOPBigMagic, false, Message.GIOPCancelRequest, CANCEL_REQ_MSG_SIZE);
         request_id = _request_id;
     }
 
     // Accessor methods
 
-    public int getRequestId() {
+    public int getRequestId()
+    {
         return this.request_id;
     }
 
     // IO methods
 
-    public void read(org.omg.CORBA.portable.InputStream istream) {
+    public void read(org.omg.CORBA.portable.InputStream istream)
+    {
         super.read(istream);
         this.request_id = istream.read_ulong();
     }
 
-    public void write(org.omg.CORBA.portable.OutputStream ostream) {
+    public void write(org.omg.CORBA.portable.OutputStream ostream)
+    {
         super.write(ostream);
         ostream.write_ulong(this.request_id);
     }
 
-    public void callback(MessageHandler handler)
-        throws java.io.IOException
+    public void callback(MessageHandler handler) throws java.io.IOException
     {
         handler.handleInput(this);
     }
