+++ src/main/java/org/jboss/com/sun/corba/se/impl/protocol/giopmsgheaders/CancelRequestMessage_1_2.java
@@ -29,47 +29,52 @@ import org.jboss.com.sun.corba.se.spi.ior.iiop.GIOPVersion;
 
 /**
  * This implements the GIOP 1.2 CancelRequest header.
- *
+ * 
  * @author Ram Jeyaraman 05/14/2000
  */
 
-public final class CancelRequestMessage_1_2 extends Message_1_1
-        implements CancelRequestMessage {
+public final class CancelRequestMessage_1_2 extends Message_1_1 implements CancelRequestMessage
+{
 
     // Instance variables
 
-    private int request_id = (int) 0;
+    private int request_id = 0;
 
     // Constructors
 
-    CancelRequestMessage_1_2() {}
+    CancelRequestMessage_1_2()
+    {
+    }
 
-    CancelRequestMessage_1_2(int _request_id) {
-        super(Message.GIOPBigMagic, GIOPVersion.V1_2, FLAG_NO_FRAG_BIG_ENDIAN,
-            Message.GIOPCancelRequest, CANCEL_REQ_MSG_SIZE);
+    CancelRequestMessage_1_2(int _request_id)
+    {
+        super(Message.GIOPBigMagic, GIOPVersion.V1_2, FLAG_NO_FRAG_BIG_ENDIAN, Message.GIOPCancelRequest,
+                CANCEL_REQ_MSG_SIZE);
         request_id = _request_id;
     }
 
     // Accessor methods
 
-    public int getRequestId() {
+    public int getRequestId()
+    {
         return this.request_id;
     }
 
     // IO methods
 
-    public void read(org.omg.CORBA.portable.InputStream istream) {
+    public void read(org.omg.CORBA.portable.InputStream istream)
+    {
         super.read(istream);
         this.request_id = istream.read_ulong();
     }
 
-    public void write(org.omg.CORBA.portable.OutputStream ostream) {
+    public void write(org.omg.CORBA.portable.OutputStream ostream)
+    {
         super.write(ostream);
         ostream.write_ulong(this.request_id);
     }
 
-    public void callback(MessageHandler handler)
-        throws java.io.IOException
+    public void callback(MessageHandler handler) throws java.io.IOException
     {
         handler.handleInput(this);
     }
